refactor: move to separate handlers, extract examples from readme and add postman collection and improvement comments into the readme
