+++ src/app.js
@@ -1,12 +1,16 @@
 const express = require("express");
 const bodyParser = require("body-parser");
-const { Op } = require("sequelize");
 
 const { pinoHttp } = require("pino-http");
 const { logger } = require("./logger");
 
 const { sequelize } = require("./model");
 const { getProfile, clientsAllowedOnly } = require("./middleware");
+
+const ContractsHandler = require("./handlers/contracts");
+const JobsHandler = require("./handlers/jobs");
+const BalancesHandler = require("./handlers/balances");
+const AdminHandler = require("./handlers/admin");
 
 const app = express();
 app.use(bodyParser.json());
@@ -15,421 +19,21 @@ app.set("sequelize", sequelize);
 app.set("models", sequelize.models);
 
 // Contracts
-
-/**
- * Get contract by id
- * @returns single contract instance by specified identifier
- */
-app.get("/contracts/:id", getProfile, async (req, res) => {
-  const { Contract } = req.app.get("models");
-  const { id } = req.params;
-  const { id: profileId } = req.profile;
-
-  try {
-    const contract = await Contract.findOne({
-      where: {
-        id,
-        [Op.or]: {
-          ClientId: profileId,
-          ContractorId: profileId,
-        },
-      },
-    });
-
-    if (!contract) return res.status(404).end();
-
-    return res.status(200).json(contract);
-  } catch (error) {
-    logger.error(error);
-    // TODO: create dedicated error and pass it to next(), propagate at the top for the top level error handler
-    return res.status(500).json({ message: "Unknown error: getting contract by id!" });
-  }
-});
-
-/**
- * List contracts
- * @returns list of contracts belonging to a user (client or contractor)
- */
-app.get("/contracts", getProfile, async (req, res) => {
-  const { Contract } = req.app.get("models");
-
-  try {
-    const contracts = await Contract.findAll({
-      where: {
-        [Op.not]: {
-          status: "terminated",
-        },
-        [Op.or]: {
-          ClientId: req.profile.id,
-          ContractorId: req.profile.id,
-        },
-      },
-    });
-
-    if (!contracts) {
-      return res.status(404).json({ message: "No contracts found!" });
-    }
-
-    return res.status(200).json(contracts);
-  } catch (error) {
-    logger.error(error);
-    return res.status(500).json({ message: "Unknown error: listing contracts!" });
-  }
-});
+app.get("/contracts/:id", getProfile, ContractsHandler.getContractById);
+app.get("/contracts", getProfile, ContractsHandler.listContracts);
 
 // Jobs
-
-/**
- * List all unpaid jobs for a user
- * @returns list all unpaid jobs for a user
- */
-app.get("/jobs/unpaid", getProfile, async (req, res) => {
-  const { Job, Contract } = req.app.get("models");
-  const { id: profileId } = req.profile;
-
-  try {
-    const contracts = await Job.findAll({
-      where: {
-        paid: null,
-      },
-      include: {
-        model: Contract,
-        where: {
-          [Op.or]: {
-            ClientId: profileId,
-            ContractorId: profileId,
-          },
-        },
-      },
-    });
-
-    if (!contracts) {
-      return res.status(404).json({ message: "No unpaid jobs found!" });
-    }
-
-    return res.status(200).json(contracts);
-  } catch (error) {
-    logger.error(error);
-    return res.status(500).json({ message: "Unknown error: listing unpaid jobs!" });
-  }
-});
-
-/**
- * Create Payment for a job
- * @returns list all unpaid jobs for a user
- */
-app.post("/jobs/:job_id/pay", getProfile, clientsAllowedOnly, async (req, res) => {
-  const { Job, Contract, Profile } = req.app.get("models");
-  const { job_id } = req.params;
-  const { id: profileId } = req.profile;
-
-  const tx = await sequelize.transaction();
-
-  try {
-    const job = await Job.findOne({
-      where: {
-        id: job_id,
-      },
-      include: {
-        model: Contract,
-        where: {
-          [Op.not]: {
-            status: "terminated",
-          },
-          ClientId: profileId,
-        },
-      },
-      transaction: tx,
-    });
-
-    if (!job) {
-      return res.status(404).json({ message: `Job with id: ${job_id} not found` });
-    }
-
-    if (job.paid) {
-      return res.status(400).json({ message: `Job with id: ${job_id} already paid` });
-    }
-
-    if (req.profile.balance < job.amount) {
-      return res.status(400).json({
-        message: "Insufficient funds in your balance!",
-        balance: req.profile.balance,
-      });
-    }
-
-    // TODO: handle decimals and potential rounding properly
-    req.profile.balance -= job.price;
-
-    await req.profile.save({ transaction: tx });
-
-    const contractorToUpdate = await Profile.findOne({
-      where: { id: job.Contract.ContractorId, type: "contractor" },
-      transaction: tx,
-    });
-
-    // TODO: handle decimals and potential rounding properly
-    contractorToUpdate.balance += job.price;
-
-    await contractorToUpdate.save({ transaction: tx });
-
-    job.paid = 1;
-    job.paymentDate = new Date();
-    await job.save({ transaction: tx });
-    await tx.commit();
-
-    // TODO: clarify and standardize the response format (DTOs needed)!
-    return res.status(200).json({ message: `Payment successful for job with id: ${job_id}` });
-  } catch (error) {
-    await tx.rollback();
-
-    logger.error(error);
-    return res.status(500).json({ message: "Unknown error: Create Job Payment!" });
-  }
-});
+app.get("/jobs/unpaid", getProfile, JobsHandler.listUnpaidJobs);
+app.post("/jobs/:job_id/pay", getProfile, clientsAllowedOnly, JobsHandler.createJobPayment);
 
 // Balances
-
-/**
- * Create Deposit for a client
- * @returns deposit request status
- */
-app.post("/balances/deposit/:userId", getProfile, async (req, res) => {
-  const { Job, Contract } = req.app.get("models");
-  const { userId } = req.params;
-  const amount = parseFloat(req.body.amount);
-
-  /**
-   * Requirement:
-   * ***POST*** `/balances/deposit/:userId`
-   * Deposits money into the the the balance of a client,
-   * a client can't deposit more than 25% his total of jobs to pay. (at the deposit moment)
-   *
-   * NOTE: A bit confused about this requirement and its wording (despite the typos and grammar)
-   * 1. Assume you can deposit to your own balance multiple times if the amount is < unpaidTotal + 25%?
-   * 2. You cannot deposit more than 25% of your total unpaid jobs at once?!?
-   */
-
-  // TODO: implement proper input validation with Joi or something similar!
-  if (!amount || amount === "NaN") {
-    return res.status(400).json({ message: "Please provide proper value for deposit amount!" });
-  }
-
-  if (parseInt(userId, 10) !== parseInt(req.profile.id, 10)) {
-    return res.status(400).json({ message: "You can only deposit to your own balance!" });
-  }
-
-  const tx = await sequelize.transaction();
-
-  try {
-    const unpaidJobs = await Job.findAll({
-      where: {
-        paid: null,
-      },
-      include: {
-        model: Contract,
-        where: {
-          status: "in_progress",
-          // [Op.not]: { // TODO: should be only jobs that are not terminated? [in_progress and new]?
-          //   status: "terminated",
-          // },
-          ClientId: userId,
-        },
-      },
-      transaction: tx,
-    });
-
-    if (!unpaidJobs || unpaidJobs.length === 0) {
-      // TODO: should we allow depositing to balance even if there are no unpaid jobs? (e.g. to pay for future jobs as credit on our account)
-      return res.status(404).json({ message: "No unpaid jobs found" });
-    }
-
-    const unpaidTotal = unpaidJobs.reduce((acc, job) => acc + job.price, 0);
-
-    if (amount > unpaidTotal * 0.25) {
-      return res.status(400).json({
-        message: "Cannot deposit more than 25% of unpaid jobs total amount!",
-        total: unpaidTotal,
-      });
-    }
-
-    req.profile.balance += amount;
-    await req.profile.save({ transaction: tx });
-
-    await tx.commit();
-
-    return res.status(200).json({ message: "Balance deposit successful" });
-  } catch (error) {
-    await tx.rollback();
-
-    logger.error(error);
-    return res.status(500).json({ message: "Unknown error: Create balance deposit!" });
-  }
-});
+app.post("/balances/deposit/:userId", getProfile, BalancesHandler.createBalanceDeposit);
 
 // Admin
+app.get("/admin/best-profession", getProfile, AdminHandler.getBestProfession);
+app.get("/admin/best-clients", getProfile, AdminHandler.listBestClients);
 
-/**
- * Admin Best Profession
- * @returns most popular profession in a given time period
- */
-app.get("/admin/best-profession", getProfile, async (req, res) => {
-  const { Job, Contract, Profile } = req.app.get("models");
-  const { start, end } = req.query;
-
-  // TODO: parse the start and end dates and validate them properly!
-  // TODO: check if start date is before end date!
-  if (!start) {
-    return res.status(400).json({
-      message: "Please provide 'start' date as query parameter! [format: YYYY-MM-DD]",
-    });
-  }
-
-  if (!end) {
-    return res.status(400).json({
-      message: "Please provide 'end' date as query parameter! [format: YYYY-MM-DD]",
-    });
-  }
-
-  try {
-    const [bestProfession] = await Job.findAll({
-      where: {
-        // TODO: change if start and end dates should be optional
-        paid: true,
-        createdAt: {
-          // TODO: clarify requirement, date between [createdAt, updatedAt or paymentDate]?
-          [Op.between]: [start, end],
-        },
-      },
-      include: {
-        model: Contract,
-        include: {
-          model: Profile,
-          as: "Contractor",
-          attributes: ["profession"],
-        },
-      },
-      group: ["profession"],
-      order: [[sequelize.fn("sum", sequelize.col("price")), "DESC"]],
-      limit: 1,
-    });
-
-    const profession = bestProfession?.Contract?.Contractor?.profession ?? null;
-
-    return res.status(200).json({ profession });
-  } catch (error) {
-    logger.error(error);
-    return res.status(500).json({ message: "Unknown error: Admin Best Profession!" });
-  }
-});
-
-/**
- * Admin Best Profession
- * @returns most popular profession in a given time period
- */
-app.get("/admin/best-profession", getProfile, async (req, res) => {
-  const { Job, Contract, Profile } = req.app.get("models");
-  const { start, end } = req.query;
-
-  // TODO: parse the start and end dates and validate them properly!
-  // TODO: check if start date is before end date!
-  if (!start) {
-    return res.status(400).json({ message: "Please provide start date as query parameter!" });
-  }
-
-  if (!end) {
-    return res.status(400).json({ message: "Please provide end date as query parameter!" });
-  }
-
-  try {
-    const [bestProfession] = await Job.findAll({
-      where: {
-        // TODO: change if start and end dates should be optional
-        paid: true,
-        createdAt: {
-          // TODO: clarify requirement, date between [createdAt, updatedAt or paymentDate]?
-          [Op.between]: [start, end],
-        },
-      },
-      include: {
-        model: Contract,
-        include: {
-          model: Profile,
-          as: "Contractor",
-          attributes: ["profession"],
-        },
-      },
-      group: ["profession"],
-      limit: 1,
-      order: [[sequelize.fn("sum", sequelize.col("price")), "DESC"]],
-    });
-
-    const profession = bestProfession?.Contract?.Contractor?.profession ?? null;
-
-    return res.status(200).json({ profession });
-  } catch (error) {
-    logger.error(error);
-    return res.status(500).json({ message: "Admin best-profession unknown error!" });
-  }
-});
-
-/**
- * Admin Best Clients
- * @returns list of clients that paid the most for jobs in the query time period
- */
-app.get("/admin/best-clients", getProfile, async (req, res) => {
-  const { Job, Contract, Profile } = req.app.get("models");
-  // eslint-disable-next-line prefer-const
-  let { start, end, limit } = req.query;
-
-  // TODO: parse the start and end dates and validate them properly!
-  // TODO: check if start date is before end date!
-  if (!start) {
-    return res.status(400).json({ message: "Please provide start date as query parameter!" });
-  }
-
-  if (!end) {
-    return res.status(400).json({ message: "Please provide end date as query parameter!" });
-  }
-
-  if (!limit || parseInt(limit, 10) === "NaN" || limit < 1) {
-    limit = 2;
-  }
-
-  try {
-    const bestClients = await Job.findAll({
-      attributes: [[sequelize.fn("sum", sequelize.col("price")), "paid"]],
-      where: {
-        // TODO: change if start and end dates should be optional
-        paid: true,
-        createdAt: {
-          // TODO: clarify requirement, date between [createdAt, updatedAt or paymentDate]?
-          [Op.between]: [start, end],
-        },
-      },
-      include: {
-        model: Contract,
-        include: {
-          model: Profile,
-          as: "Client",
-          attributes: ["id", [sequelize.literal("firstName || ' ' || lastName"), "fullName"]],
-        },
-      },
-      group: ["Contract.Client.id"],
-      order: [[sequelize.fn("sum", sequelize.col("price")), "DESC"]],
-      limit,
-    });
-
-    const response = bestClients.map((client) => ({
-      id: client.Contract.Client.id,
-      fullName: client.Contract.Client.dataValues.fullName,
-      paid: client.paid,
-    }));
-
-    return res.status(200).json(response);
-  } catch (error) {
-    logger.error(error);
-    return res.status(500).json({ message: "Unknown error: Admin Best Clients!" });
-  }
-});
+// Global error handler
+// TODO: add global error handler
 
 module.exports = app;
